refactor: split MCP OAuth and manager modules
Structure-only split of the two files left over the 500-line cap by #734.
No logic, behavior, interface or error-handling change; every moved
definition is byte-identical to its pre-split source.

src/agency_swarm/mcp/oauth.py (958 -> 477)
- oauth_flow.py (320): interactive browser/stdin callback capture
- oauth_config.py (181): MCPServerOAuth
- oauth_storage_hooks.py (53): OAuthStorageHooks
- oauth.py keeps the security-critical chain: contextvars, token callback
  registry, FileTokenStorage and create_oauth_provider

src/agency_swarm/tools/mcp_manager.py (1093 -> 218)
- mcp_persistence.py (446): PersistentMCPServerManager
- mcp_oauth_bridge.py (289): optional OAuth imports, persistence keys,
  cache-dir and handler resolution
- hosted_mcp_activation.py (181): deferred hosted MCP OAuth activation
- mcp_loop_proxy.py (75): LoopAffineAsyncProxy
- mcp_manager.py keeps default_mcp_manager and the agent/agency entry points

Both original modules re-export every name they used to own, with object
identity, so existing imports and agency_swarm/mcp/__init__.py are unchanged.
Monkeypatch targets are repointed to the new defining modules, because
module-global resolution happens where the definition lives and re-exports
do not redirect it.
