+++ src/agency_swarm/mcp/oauth.py
@@ -2,36 +2,44 @@
 
 Provides OAuth 2.0 with PKCE support for MCP servers using the MCP Python SDK's OAuthClientProvider.
 Supports unified local/SaaS token storage via RunHooks and contextvars.
+
+This module owns the security-critical storage chain: the hooks set the user contextvar,
+storage reads it per operation, bucket containment is checked, token files are written
+owner-only, and callback keys include the user segment. The interactive browser/stdin
+flow lives in :mod:`agency_swarm.mcp.oauth_flow`, the server configuration model in
+:mod:`agency_swarm.mcp.oauth_config`, and the run hooks in
+:mod:`agency_swarm.mcp.oauth_storage_hooks`; all of them are re-exported here so
+``from agency_swarm.mcp.oauth import X`` keeps working for every previously exported name.
 """
 
-import asyncio
 import contextlib
 import json
 import logging
 import os
-import select
-import sys
 import tempfile
 import uuid
-import webbrowser
-from collections.abc import Awaitable, Callable, Coroutine
+from collections.abc import Awaitable, Callable
 from contextvars import ContextVar
 from dataclasses import dataclass, field
-from html import escape
 from pathlib import Path
-from typing import Any, ClassVar, Literal, TypedDict, cast
-from urllib.parse import parse_qs, urlparse, urlsplit
+from typing import Literal, TypedDict, cast
 
-from agents import RunHooks
 from mcp.client.auth import OAuthClientProvider
-from mcp.client.auth.oauth2 import TokenStorage
 from mcp.shared.auth import (
     OAuthClientInformationFull,
-    OAuthClientMetadata,
     OAuthToken,
 )
-from pydantic import AnyUrl
 
+from .oauth_flow import (
+    _can_poll_stdin_for_callback,  # noqa: F401
+    _can_use_local_callback_server,  # noqa: F401
+    _listen_for_callback_once,  # noqa: F401
+    _parse_callback_response,  # noqa: F401
+    _prompt_for_callback_url,  # noqa: F401
+    _prompt_for_callback_url_polling,  # noqa: F401
+    default_callback_handler,
+    default_redirect_handler,
+)
 from .oauth_user import build_oauth_cache_segment, build_oauth_user_segment
 
 # Contextvar for per-user token isolation
@@ -386,501 +394,12 @@ class FileTokenStorage:
                 legacy_file.unlink()
 
 
-class OAuthStorageHooks(RunHooks):  # type: ignore[type-arg]
-    """RunHooks implementation for OAuth token storage.
-
-    Sets the user_id contextvar from MasterContext at the start of each run,
-    enabling per-user token isolation in FileTokenStorage.
-
-    Usage:
-        from agency_swarm import Agency
-        from agency_swarm.mcp.oauth import OAuthStorageHooks
-
-        agency = Agency(
-            [agent],
-            oauth_token_path="./data",
-            user_context={"user_id": "user_123"},
-            hooks=[OAuthStorageHooks()],
-        )
-    """
-
-    def on_run_start(self, *, context: Any, **kwargs: Any) -> None:
-        """Set user_id contextvar from MasterContext at run start."""
-        user_id = context.user_context.get("user_id") if hasattr(context, "user_context") else None
-        _user_id_context.set(user_id)
-        logger.debug(f"OAuth user_id context set to: {user_id}")
-
-    def on_run_end(self, *, context: Any, result: Any, **kwargs: Any) -> None:
-        """Clear user_id contextvar at run end."""
-        _user_id_context.set(None)
-        logger.debug("OAuth user_id context cleared")
-
-    async def on_agent_start(self, context: Any, agent: Any) -> None:
-        """Bridge the current Agents SDK hook to the legacy run-start behavior."""
-        self.on_run_start(context=context.context)
-
-    async def on_agent_end(self, context: Any, agent: Any, output: Any) -> None:
-        """Bridge the current Agents SDK hook to the legacy run-end behavior."""
-        self.on_run_end(context=context.context, result=output)
-
-
-@dataclass(eq=False)
-class MCPServerOAuth:
-    """Configuration for an OAuth-enabled MCP server.
-
-    Attributes:
-        url: MCP server URL
-        name: Unique identifier for this server
-        client_id: OAuth client ID (reads from env if None and use_env_credentials=True)
-        client_secret: OAuth client secret (reads from env if None and use_env_credentials=True)
-        scopes: List of OAuth scopes to request
-        redirect_uri: OAuth redirect URI for callback
-        cache_dir: Directory for token storage (uses default if None)
-        storage: Custom token storage implementation (overrides cache_dir)
-        storage_factory: Factory function to create storage per-request (for multi-tenant)
-        client_metadata: Full OAuth client metadata (overrides simple params)
-        auth_server_url: Base URL for OAuth discovery when different from MCP endpoint
-        use_env_credentials: If False, don't read client_id/secret from environment.
-            Set to False for self-hosted servers using Dynamic Client Registration (DCR).
-        redirect_handler: Custom handler for OAuth redirect (opens browser by default)
-        callback_handler: Custom handler to receive OAuth callback code
-    """
-
-    DEFAULT_REDIRECT_URI: ClassVar[str] = "http://localhost:8000/auth/callback"
-
-    url: str
-    name: str
-    client_id: str | None = None
-    client_secret: str | None = None
-    scopes: list[str] = field(default_factory=lambda: ["user"])
-    redirect_uri: str | None = None
-    cache_dir: Path | None = None
-    storage: TokenStorage | None = None
-    storage_factory: Callable[[str, str], TokenStorage] | None = None
-    client_metadata: OAuthClientMetadata | None = None
-    auth_server_url: str | None = None
-    use_env_credentials: bool = True
-    redirect_handler: OAuthRedirectHandler | None = None
-    callback_handler: OAuthCallbackHandler | None = None
-
-    def _resolve_client_id(self) -> str | None:
-        """Return the client_id if provided explicitly or via environment."""
-        if self.client_id:
-            return self.client_id
-
-        if not self.use_env_credentials:
-            return None
-
-        env_var = f"{self.name.upper().replace('-', '_')}_CLIENT_ID"
-        return os.getenv(env_var)
-
-    def get_client_id(self) -> str:
-        """Get client ID from config or environment."""
-        client_id = self._resolve_client_id()
-        if client_id:
-            return client_id
-
-        env_var = f"{self.name.upper().replace('-', '_')}_CLIENT_ID"
-        raise ValueError(
-            f"No client_id provided for {self.name}. Set {env_var} environment variable or pass client_id parameter."
-        )
-
-    def get_client_secret(self) -> str | None:
-        """Get client secret from config or environment."""
-        if self.client_secret:
-            return self.client_secret
-
-        if not self.use_env_credentials:
-            return None
-
-        # Try server-specific env var
-        env_var = f"{self.name.upper().replace('-', '_')}_CLIENT_SECRET"
-        return os.getenv(env_var)
-
-    def get_cache_dir(self) -> Path:
-        """Get cache directory for token storage."""
-        if self.cache_dir:
-            return self.cache_dir
-        return get_default_cache_dir()
-
-    def build_client_metadata(self) -> OAuthClientMetadata:
-        """Build effective OAuth client metadata from config."""
-        if self.client_metadata:
-            metadata = self.client_metadata.model_copy(deep=True)
-        else:
-            metadata = OAuthClientMetadata(
-                client_name=f"Agency Swarm - {self.name}",
-                redirect_uris=[AnyUrl(self.get_redirect_uri())],
-                grant_types=["authorization_code", "refresh_token"],
-                response_types=["code"],
-                scope=" ".join(self.scopes),
-            )
-
-        if self.get_client_secret() and metadata.token_endpoint_auth_method is None:
-            return metadata.model_copy(update={"token_endpoint_auth_method": "client_secret_basic"})
-        return metadata
-
-    def get_client_id_optional(self) -> str | None:
-        """Return the resolved client_id without raising."""
-        return self._resolve_client_id()
-
-    def get_redirect_uri(self) -> str:
-        """Resolve redirect URI with explicit > server env > global env > default."""
-        if self.redirect_uri:
-            return self.redirect_uri
-        server_env = f"{self.name.upper().replace('-', '_')}_REDIRECT_URI"
-        if os.getenv(server_env):
-            return os.getenv(server_env, self.DEFAULT_REDIRECT_URI)
-        if os.getenv("OAUTH_CALLBACK_URL"):
-            return os.getenv("OAUTH_CALLBACK_URL", self.DEFAULT_REDIRECT_URI)
-        return self.DEFAULT_REDIRECT_URI
-
-    def get_callback_redirect_uri(self, client_metadata: OAuthClientMetadata) -> str:
-        """Return the advertised URI used by the default local callback listener."""
-        if client_metadata.redirect_uris:
-            return str(client_metadata.redirect_uris[0])
-        return self.get_redirect_uri()
-
-    def build_client_information(self) -> OAuthClientInformationFull | None:
-        """Return prepopulated client information when static credentials exist."""
-        client_id = self.get_client_id_optional()
-        if not client_id:
-            return None
-
-        client_secret = self.get_client_secret()
-        metadata = self.build_client_metadata()
-        metadata_data = metadata.model_dump(by_alias=True, exclude_none=True)
-        metadata_data.update(
-            client_id=client_id,
-            client_secret=client_secret,
-        )
-
-        return OAuthClientInformationFull(**metadata_data)
-
-    def get_client_identity(self) -> str:
-        """Return a stable, secret-free identity for OAuth persistence."""
-        metadata = self.build_client_metadata()
-        identity = json.dumps(
-            {
-                "auth_server_url": self.get_auth_server_url(),
-                "client_id": self.get_client_id_optional(),
-                "client_metadata": metadata.model_dump(mode="json", by_alias=True, exclude_none=True),
-            },
-            sort_keys=True,
-            separators=(",", ":"),
-        )
-        return build_oauth_cache_segment(identity, max_prefix_length=48)
-
-    def get_auth_server_url(self) -> str | None:
-        """Return the OAuth authorization server base URL."""
-        if self.auth_server_url:
-            return self.auth_server_url
-
-        parsed = urlsplit(self.url)
-        if not parsed.scheme or not parsed.netloc:
-            return None
-        return f"{parsed.scheme}://{parsed.netloc}"
-
-
-async def default_redirect_handler(auth_url: str) -> None:
-    """Default handler for OAuth redirect - opens browser.
-
-    Args:
-        auth_url: Authorization URL to visit
-    """
-    print(f"\n{'=' * 80}")
-    print("OAuth Authentication Required")
-    print(f"{'=' * 80}")
-    print(f"\nOpening browser for authentication: {auth_url}\n")
-    print("If the browser doesn't open automatically, please visit the URL above.")
-    print("The terminal will automatically capture the callback or you can paste it manually.")
-    print(f"{'=' * 80}\n")
-
-    try:
-        webbrowser.open(auth_url)
-    except Exception:
-        logger.exception("Failed to open browser")
-
-
-def _parse_callback_response(callback_url: str) -> tuple[str, str | None]:
-    """Parse authorization code and state from a callback URL."""
-    parsed = urlparse(callback_url)
-    params = parse_qs(parsed.query)
-
-    if "error" in params:
-        error = params["error"][0]
-        error_description = params.get("error_description", ["Unknown error"])[0]
-        raise ValueError(f"OAuth error: {error} - {error_description}")
-
-    if "code" not in params:
-        raise ValueError("No authorization code found in callback URL")
-
-    code = params["code"][0]
-    state = params.get("state", [None])[0]
-    return code, state
-
-
-async def _prompt_for_callback_url() -> tuple[str, str | None]:
-    """Prompt the user to paste the callback URL."""
-    print("\nAfter authorizing, you will be redirected to a callback URL.")
-    print("Paste the full URL here if automatic capture does not complete.\n")
-
-    loop = asyncio.get_event_loop()
-    callback_url = await loop.run_in_executor(None, lambda: input("Callback URL: ").strip())
-    return _parse_callback_response(callback_url)
-
-
-def _can_poll_stdin_for_callback() -> bool:
-    """Return True when stdin supports polling-based callback input."""
-    stdin = sys.stdin
-    if stdin is None or not hasattr(stdin, "isatty") or not stdin.isatty():
-        return False
-    if os.name == "nt":
-        try:
-            import msvcrt  # type: ignore[attr-defined]
-
-            return callable(getattr(msvcrt, "kbhit", None)) and callable(getattr(msvcrt, "getwche", None))
-        except ImportError:
-            return False
-    try:
-        select.select([stdin], [], [], 0)
-    except (OSError, ValueError, TypeError):
-        return False
-    return True
-
-
-async def _prompt_for_callback_url_polling(poll_interval: float = 0.5) -> tuple[str, str | None]:
-    """Prompt for callback URL using poll-based stdin reads that cancel cleanly."""
-    print("\nAfter authorizing, you will be redirected to a callback URL.")
-    print("Paste the full URL here if automatic capture does not complete.\n")
-
-    stdin = sys.stdin
-    if stdin is None:
-        raise EOFError("stdin is unavailable")
-
-    loop = asyncio.get_event_loop()
-    if os.name == "nt":
-        try:
-            import msvcrt  # type: ignore[attr-defined]
-        except ImportError as exc:  # pragma: no cover - platform-specific fallback
-            raise EOFError("stdin polling backend unavailable") from exc
-        kbhit = getattr(msvcrt, "kbhit", None)
-        getwche = getattr(msvcrt, "getwche", None)
-        if not callable(kbhit) or not callable(getwche):
-            raise EOFError("stdin polling backend unavailable")
-
-        line_buffer = ""
-        while True:
-            while kbhit():
-                ch = getwche()
-                if ch in {"\r", "\n"}:
-                    print("")
-                    callback_url = line_buffer.strip()
-                    line_buffer = ""
-                    if callback_url == "":
-                        break
-                    return _parse_callback_response(callback_url)
-                if ch == "\x08":
-                    line_buffer = line_buffer[:-1]
-                    continue
-                line_buffer += ch
-            await asyncio.sleep(poll_interval)
-
-    while True:
-
-        def _wait_for_input() -> bool:
-            ready, _, _ = select.select([stdin], [], [], poll_interval)
-            return bool(ready)
-
-        try:
-            has_input = await loop.run_in_executor(None, _wait_for_input)
-        except (OSError, ValueError) as exc:
-            raise EOFError("stdin does not support polling") from exc
-
-        if not has_input:
-            continue
-
-        callback_url = stdin.readline()
-        if callback_url == "":
-            raise EOFError("stdin reached EOF")
-        callback_url = callback_url.strip()
-        if callback_url == "":
-            continue
-        return _parse_callback_response(callback_url)
-
-
-def _can_use_local_callback_server(redirect_uri: str) -> bool:
-    """Return True if we can bind a local HTTP server for the redirect URI."""
-    parsed = urlparse(redirect_uri)
-    if parsed.scheme != "http":
-        return False
-
-    host = parsed.hostname or ""
-    return host in {"localhost", "127.0.0.1"}
-
-
-async def _listen_for_callback_once(redirect_uri: str, timeout: float = 300.0) -> tuple[str, str | None]:
-    """Start a local HTTP server and capture the first callback request."""
-    parsed = urlparse(redirect_uri)
-    host = parsed.hostname or "localhost"
-    port = parsed.port or 80
-    path = parsed.path or "/auth/callback"
-    loop = asyncio.get_event_loop()
-    result: asyncio.Future[tuple[str, str | None]] = loop.create_future()
-
-    async def _handle_connection(reader: asyncio.StreamReader, writer: asyncio.StreamWriter) -> None:
-        try:
-            data = await reader.readuntil(b"\r\n\r\n")
-        except asyncio.IncompleteReadError:
-            writer.close()
-            await writer.wait_closed()
-            return
-
-        request_line = data.split(b"\r\n", 1)[0].decode(errors="ignore")
-        parts = request_line.split(" ")
-        target = parts[1] if len(parts) >= 2 else ""
-        target_parsed = urlparse(target)
-        # Some browsers send absolute URLs, others only the path.
-        request_path = target_parsed.path or target
-
-        status_line = "HTTP/1.1 200 OK\r\n"
-        body = (
-            "<html><body>"
-            "<h1>You may close this tab.</h1>"
-            "<p>Return to Agency Swarm. The terminal captured your authorization code.</p>"
-            "</body></html>"
-        )
-
-        try:
-            if request_path != path:
-                status_line = "HTTP/1.1 404 Not Found\r\n"
-                body = "<html><body><h1>404 Not Found</h1></body></html>"
-                writer.write(
-                    f"{status_line}Content-Type: text/html\r\nContent-Length: {len(body)}\r\n\r\n{body}".encode()
-                )
-                await writer.drain()
-                writer.close()
-                await writer.wait_closed()
-                return
-
-            params = parse_qs(target_parsed.query)
-
-            # Handle OAuth provider error responses (e.g., user denied authorization)
-            if "error" in params:
-                error = params["error"][0]
-                error_description = params.get("error_description", ["Unknown error"])[0]
-                status_line = "HTTP/1.1 400 Bad Request\r\n"
-                body = (
-                    f"<html><body><h1>OAuth Error</h1><p>{escape(error)}: {escape(error_description)}</p></body></html>"
-                )
-                body_bytes = body.encode()
-                headers = f"{status_line}Content-Type: text/html\r\nContent-Length: {len(body_bytes)}\r\n\r\n".encode()
-                writer.write(headers + body_bytes)
-                await writer.drain()
-                writer.close()
-                await writer.wait_closed()
-                if not result.done():
-                    result.set_exception(ValueError(f"OAuth error: {error} - {error_description}"))
-                return
-
-            if "code" not in params:
-                status_line = "HTTP/1.1 400 Bad Request\r\n"
-                body = "<html><body><h1>Missing code parameter.</h1></body></html>"
-                writer.write(
-                    f"{status_line}Content-Type: text/html\r\nContent-Length: {len(body)}\r\n\r\n{body}".encode()
-                )
-                await writer.drain()
-                writer.close()
-                await writer.wait_closed()
-                return
-
-            if not result.done():
-                code = params["code"][0]
-                state = params.get("state", [None])[0]
-                result.set_result((code, state))
-
-            writer.write(f"{status_line}Content-Type: text/html\r\nContent-Length: {len(body)}\r\n\r\n{body}".encode())
-            await writer.drain()
-        finally:
-            writer.close()
-            await writer.wait_closed()
-
-    server = await asyncio.start_server(_handle_connection, host=host, port=port)
-    print(f"\nListening for OAuth callback at {redirect_uri}\n")
-
-    try:
-        return await asyncio.wait_for(result, timeout=timeout)
-    finally:
-        server.close()
-        await server.wait_closed()
-
-
-async def default_callback_handler(redirect_uri: str | None = None, timeout: float = 300.0) -> tuple[str, str | None]:
-    """Default handler for OAuth callback.
-
-    Tries to capture the callback automatically using a local HTTP server and falls
-    back to a manual prompt when automatic capture is not possible.
-
-    Returns:
-        Tuple of (authorization_code, state)
-    """
-    redirect_target = redirect_uri or "http://localhost:8000/auth/callback"
-
-    if _can_use_local_callback_server(redirect_target):
-        listener_task = asyncio.create_task(_listen_for_callback_once(redirect_target, timeout=timeout))
-        prompt_factory: Callable[[], Coroutine[object, object, tuple[str, str | None]]] | None = None
-        if _can_poll_stdin_for_callback():
-            prompt_factory = _prompt_for_callback_url_polling
-        prompt_task: asyncio.Task[tuple[str, str | None]] | None = None
-        if prompt_factory is not None:
-            prompt_task = asyncio.create_task(prompt_factory())
-        tasks: list[asyncio.Task[tuple[str, str | None]]] = [listener_task]
-        if prompt_task is not None:
-            tasks.append(prompt_task)
-        try:
-            if prompt_task is None:
-                return await listener_task
-
-            while True:
-                done, _ = await asyncio.wait(set(tasks), return_when=asyncio.FIRST_COMPLETED)
-                if prompt_task in done:
-                    try:
-                        return prompt_task.result()
-                    except EOFError:
-                        if listener_task in done:
-                            return listener_task.result()
-                        return await listener_task
-                    except ValueError as exc:
-                        print(f"Invalid callback URL: {exc}. Please try again.")
-                        if listener_task in done:
-                            return listener_task.result()
-                        if prompt_factory is None:
-                            return await listener_task
-                        prompt_task = asyncio.create_task(prompt_factory())
-                        tasks = [listener_task, prompt_task]
-                        continue
-                if listener_task in done:
-                    return listener_task.result()
-        except OSError as exc:
-            logger.warning("Local callback server unavailable (%s); falling back to manual entry.", exc)
-        except TimeoutError:
-            logger.warning("Timed out waiting for local OAuth callback; falling back to manual entry.")
-        finally:
-            for task in tasks:
-                if not task.done():
-                    task.cancel()
-                    with contextlib.suppress(asyncio.CancelledError):
-                        await task
-
-    try:
-        return await _prompt_for_callback_url()
-    except EOFError as exc:
-        # Non-interactive environments (e.g. background workers) cannot
-        # satisfy the fallback stdin prompt.
-        raise RuntimeError(
-            "OAuth callback input is unavailable in non-interactive mode. "
-            "Use FastAPI OAuth handlers or provide a callback URL."
-        ) from exc
+# Compatibility re-exports, deferred to here on purpose: both modules read the
+# contextvar and cache helpers defined above, so they can only be imported once those
+# names exist. This keeps ``from agency_swarm.mcp.oauth import X`` working for every name
+# this module used to own.
+from .oauth_config import MCPServerOAuth  # noqa: E402
+from .oauth_storage_hooks import OAuthStorageHooks  # noqa: E402, F401
 
 
 async def create_oauth_provider(
