+++ src/agency_swarm/tools/mcp_manager.py
@@ -1,944 +1,69 @@
-import asyncio
-import copy
-import inspect
-import logging
-import threading
-from collections.abc import Awaitable, Callable, Coroutine
-from concurrent.futures import Future
+"""Entry points that bind an agent's or agency's MCP servers to the persistent manager.
+
+The registry itself lives in :mod:`agency_swarm.tools.mcp_persistence`, the loop-affine
+proxy in :mod:`agency_swarm.tools.mcp_loop_proxy`, the optional OAuth wiring in
+:mod:`agency_swarm.tools.mcp_oauth_bridge`, and hosted MCP OAuth activation in
+:mod:`agency_swarm.tools.hosted_mcp_activation`. All of their names are re-exported here
+so ``from agency_swarm.tools.mcp_manager import X`` keeps working unchanged.
+"""
+
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, TypedDict, cast
+from typing import TYPE_CHECKING, Any
 
 from agents import FunctionTool
 
-from agency_swarm.mcp.oauth_user import build_oauth_user_segment
+from agency_swarm.mcp.oauth_user import build_oauth_user_segment  # noqa: F401
+from agency_swarm.tools.hosted_mcp_activation import (
+    _HOSTED_MCP_OAUTH_ACTIVATION_HANDLER_ATTR,  # noqa: F401
+    _HOSTED_MCP_OAUTH_ORIGINAL_TOOL_ATTR,  # noqa: F401
+    _HOSTED_MCP_OAUTH_ORIGINAL_TOOLS_ATTR,  # noqa: F401
+    _authorize_hosted_mcp_tool,  # noqa: F401
+    _HostedMCPTool,  # noqa: F401
+    _stage_deferred_hosted_mcp_tools,
+    restore_hosted_mcp_oauth_tools,  # noqa: F401
+)
 from agency_swarm.tools.hosted_mcp_oauth import (
-    enable_hosted_mcp_tool_oauth,
-    is_hosted_mcp_tool_oauth_enabled,
+    enable_hosted_mcp_tool_oauth,  # noqa: F401
+    is_hosted_mcp_tool_oauth_enabled,  # noqa: F401
+)
+from agency_swarm.tools.mcp_loop_proxy import LoopAffineAsyncProxy
+from agency_swarm.tools.mcp_oauth_bridge import (
+    _MANAGED_OAUTH_CACHE_DIR_ATTR,  # noqa: F401
+    _OAUTH_AVAILABLE,
+    OAuthCallbackHandler,  # noqa: F401
+    OAuthHandlerMap,  # noqa: F401
+    OAuthRedirectHandler,  # noqa: F401
+    _build_persistence_key,
+    _clone_oauth_candidate,
+    _create_oauth_provider,  # noqa: F401
+    _current_oauth_handlers,  # noqa: F401
+    _get_oauth_runtime_context,
+    _get_oauth_user_id,
+    _MCPServerOAuth,  # noqa: F401
+    _MCPServerOAuthClient,  # noqa: F401
+    _oauth_endpoint_key_segment,  # noqa: F401
+    _oauth_request_key_suffix,
+    _oauth_store_key_segment,  # noqa: F401
+    _process_oauth_servers,
+    _sanitize_oauth_registry_user_id,  # noqa: F401
+    _set_oauth_runtime_context,  # noqa: F401
+    _set_oauth_user_id,  # noqa: F401
+    _sync_oauth_client_handlers,
+    apply_managed_oauth_cache_dir,  # noqa: F401
+    bind_managed_oauth_cache_dir,  # noqa: F401
+    get_active_oauth_user_id,  # noqa: F401
+)
+from agency_swarm.tools.mcp_persistence import (
+    _OAUTH_LIST_TOOLS_TIMEOUT_GRACE_SECONDS,  # noqa: F401
+    _OAUTH_LIST_TOOLS_TIMEOUT_SECONDS,  # noqa: F401
+    PersistentMCPServerManager,
 )
 
-_HostedMCPTool: type[Any] | None
-try:
-    from agents.tool import HostedMCPTool as _HostedMCPTool
-except ImportError:  # pragma: no cover - compatibility with older Agents SDKs
-    _HostedMCPTool = None
-
 if TYPE_CHECKING:
-    from mcp.client.auth import OAuthClientProvider
-
     from agency_swarm.agent.core import Agent
-    from agency_swarm.mcp.oauth import MCPServerOAuth
-
-logger = logging.getLogger(__name__)
-
-OAuthRedirectHandler = Callable[[str], Awaitable[None]]
-OAuthCallbackHandler = Callable[[], Awaitable[tuple[str, str | None]]]
-
-
-class OAuthHandlerMap(TypedDict, total=False):
-    redirect: OAuthRedirectHandler
-    callback: OAuthCallbackHandler
-
-
-# OAuth support - imported conditionally to avoid circular imports
-_OAUTH_AVAILABLE = False
-_MCPServerOAuth: type | None = None
-_MCPServerOAuthClient: type | None = None
-_create_oauth_provider: Callable[..., Coroutine[object, object, "OAuthClientProvider"]] | None = None
-_set_oauth_user_id: Callable[[str | None], None] | None = None
-_get_oauth_user_id: Callable[[], str | None] | None = None
-_set_oauth_runtime_context: Callable[[Any | None], None] | None = None
-_get_oauth_runtime_context: Callable[[], Any | None] | None = None
-
-try:
-    from agency_swarm.mcp.oauth import (
-        MCPServerOAuth as _MCPServerOAuth_impl,
-        create_oauth_provider as _create_oauth_provider_impl,
-        get_oauth_runtime_context as _get_oauth_runtime_context_impl,
-        get_oauth_user_id as _get_oauth_user_id_impl,
-        set_oauth_runtime_context as _set_oauth_runtime_context_impl,
-        set_oauth_user_id as _set_oauth_user_id_impl,
-    )
-    from agency_swarm.mcp.oauth_client import MCPServerOAuthClient as _MCPServerOAuthClient_impl
-
-    _MCPServerOAuth = _MCPServerOAuth_impl
-    _MCPServerOAuthClient = _MCPServerOAuthClient_impl
-    _create_oauth_provider = _create_oauth_provider_impl
-    _set_oauth_user_id = _set_oauth_user_id_impl
-    _get_oauth_user_id = _get_oauth_user_id_impl
-    _set_oauth_runtime_context = _set_oauth_runtime_context_impl
-    _get_oauth_runtime_context = _get_oauth_runtime_context_impl
-    _OAUTH_AVAILABLE = True
-except ImportError:
-    logger.debug("OAuth support not available - install MCP SDK to enable")
-
-_HOSTED_MCP_OAUTH_ORIGINAL_TOOL_ATTR = "_agency_swarm_hosted_mcp_oauth_original_tool"
-_HOSTED_MCP_OAUTH_ORIGINAL_TOOLS_ATTR = "_agency_swarm_hosted_mcp_oauth_original_tools"
-_HOSTED_MCP_OAUTH_ACTIVATION_HANDLER_ATTR = "_agency_swarm_hosted_mcp_oauth_activation_handler"
-_MANAGED_OAUTH_CACHE_DIR_ATTR = "_agency_swarm_managed_oauth_cache_dir"
-
-
-def get_active_oauth_user_id() -> str | None:
-    """Return the OAuth user whose token bucket the current context is bound to."""
-    if _get_oauth_user_id is None:
-        return None
-    return _get_oauth_user_id()
-
-
-def _sanitize_oauth_registry_user_id(user_id: str) -> str:
-    return build_oauth_user_segment(user_id, max_prefix_length=96)
-
-
-def _oauth_store_key_segment(config: Any) -> str:
-    storage = getattr(config, "storage", None)
-    if storage is not None:
-        return _sanitize_oauth_registry_user_id(f"storage:{id(storage)}")
-
-    storage_factory = getattr(config, "storage_factory", None)
-    if storage_factory is not None:
-        return _sanitize_oauth_registry_user_id(f"storage_factory:{id(storage_factory)}")
-
-    cache_dir = getattr(config, "cache_dir", None)
-    if cache_dir is not None:
-        path = Path(cache_dir).expanduser().resolve(strict=False)
-        return _sanitize_oauth_registry_user_id(f"cache:{path}")
-
-    return "default"
-
-
-def _oauth_endpoint_key_segment(config: Any) -> str:
-    url = getattr(config, "url", None)
-    if not isinstance(url, str) or url == "":
-        return "unknown-url"
-    return _sanitize_oauth_registry_user_id(f"url:{url}")
-
-
-def _oauth_request_key_suffix(request_id: str) -> str:
-    return f"::{_sanitize_oauth_registry_user_id(f'request:{request_id}')}"
-
-
-def _build_persistence_key(server: Any, oauth_user_id: str | None) -> str:
-    """Build process-level persistence key for MCP servers.
-
-    OAuth clients are keyed by server, endpoint, OAuth client identity, user,
-    and token store to avoid cross-client, cross-user, or cross-agency reuse.
-    Non-OAuth servers keep name-only keys.
-    """
-    actual = getattr(server, "_server", server)
-    name = getattr(actual, "name", None)
-    if not isinstance(name, str) or name == "":
-        return ""
-    if _MCPServerOAuthClient is None or not isinstance(actual, _MCPServerOAuthClient):
-        return name
-    if not isinstance(oauth_user_id, str) or oauth_user_id == "":
-        user_segment = "default"
-    else:
-        user_segment = _sanitize_oauth_registry_user_id(oauth_user_id)
-    oauth_client = cast(Any, actual)
-    endpoint_segment = _oauth_endpoint_key_segment(oauth_client.oauth_config)
-    client_segment = _sanitize_oauth_registry_user_id(f"client:{oauth_client.oauth_config.get_client_identity()}")
-    store_segment = _oauth_store_key_segment(oauth_client.oauth_config)
-    runtime_context = _get_oauth_runtime_context() if _get_oauth_runtime_context is not None else None
-    runtime_request_id = getattr(runtime_context, "request_id", None) if runtime_context is not None else None
-    if getattr(runtime_context, "mode", None) == "saas_stream" and isinstance(runtime_request_id, str):
-        key_prefix = f"{name}::oauth::{endpoint_segment}::{client_segment}::{user_segment}::{store_segment}"
-        return f"{key_prefix}{_oauth_request_key_suffix(runtime_request_id)}"
-    return f"{name}::oauth::{endpoint_segment}::{client_segment}::{user_segment}::{store_segment}"
-
-
-def apply_managed_oauth_cache_dir(config: Any, cache_dir: Path | None) -> None:
-    """Apply an Agency-managed OAuth cache dir without overwriting explicit server config."""
-    if cache_dir is None:
-        return
-    if getattr(config, "cache_dir", None) is None or bool(getattr(config, _MANAGED_OAUTH_CACHE_DIR_ATTR, False)):
-        config.cache_dir = cache_dir
-        setattr(config, _MANAGED_OAUTH_CACHE_DIR_ATTR, True)
-
-
-def bind_managed_oauth_cache_dir(config: Any, cache_dir: Path | None) -> Any:  # noqa: ANN401
-    """Bind an Agency-managed cache dir without retargeting another Agency's config."""
-    if cache_dir is None:
-        return config
-    current = getattr(config, "cache_dir", None)
-    is_managed = bool(getattr(config, _MANAGED_OAUTH_CACHE_DIR_ATTR, False))
-    if current is not None and not is_managed:
-        return config
-    if is_managed and current != cache_dir:
-        config = copy.copy(config)
-    apply_managed_oauth_cache_dir(config, cache_dir)
-    return config
-
-
-def _clone_oauth_candidate(server: Any) -> Any:
-    """Return a fresh OAuth client when the current object is already user-bound."""
-    if not _OAUTH_AVAILABLE or _MCPServerOAuthClient is None:
-        return server
-
-    actual = getattr(server, "_server", server)
-    if not isinstance(actual, _MCPServerOAuthClient):
-        return server
-    client = cast(Any, actual)
-
-    runtime_context = _get_oauth_runtime_context() if _get_oauth_runtime_context is not None else None
-    preserve_client_handlers = runtime_context is None or (
-        client.oauth_config.redirect_handler is not None or client.oauth_config.callback_handler is not None
-    )
-    handlers: OAuthHandlerMap = {}
-    if preserve_client_handlers and client._redirect_handler is not None:
-        handlers["redirect"] = client._redirect_handler
-    if preserve_client_handlers and client._callback_handler is not None:
-        handlers["callback"] = client._callback_handler
-    return _MCPServerOAuthClient(client.oauth_config, handlers or None)
-
-
-def _current_oauth_handlers(client: Any) -> OAuthHandlerMap:
-    """Return the handlers that should be active for the current OAuth request."""
-    handlers: OAuthHandlerMap = {}
-    if client._redirect_handler is not None:
-        handlers["redirect"] = client._redirect_handler
-    if client._callback_handler is not None:
-        handlers["callback"] = client._callback_handler
-
-    runtime_context = _get_oauth_runtime_context() if _get_oauth_runtime_context is not None else None
-    if runtime_context is not None:
-        if "redirect" not in handlers and runtime_context.redirect_handler_factory is not None:
-            handlers["redirect"] = runtime_context.redirect_handler_factory(client.name)
-        if "callback" not in handlers and runtime_context.callback_handler_factory is not None:
-            handlers["callback"] = runtime_context.callback_handler_factory(client.name)
-
-    if "redirect" not in handlers and client.oauth_config.redirect_handler is not None:
-        handlers["redirect"] = client.oauth_config.redirect_handler
-    if "callback" not in handlers and client.oauth_config.callback_handler is not None:
-        handlers["callback"] = client.oauth_config.callback_handler
-    return handlers
-
-
-_OAUTH_LIST_TOOLS_TIMEOUT_SECONDS = 620.0
-_OAUTH_LIST_TOOLS_TIMEOUT_GRACE_SECONDS = 20.0
-
-
-class PersistentMCPServerManager:
-    """Process-level registry for MCP servers with persistent connections.
-    Servers are keyed by their readable `name` attribute. New agencies/agents
-    should reuse these instances instead of creating new ones to preserve a
-    single connection per process.
-    """
-
-    def __init__(self) -> None:
-        self._servers: dict[str, Any] = {}
-        self._lock: asyncio.Lock = asyncio.Lock()
-        self._bg_loop: asyncio.AbstractEventLoop | None = None
-        self._bg_thread: threading.Thread | None = None
-        self._sync_shutdown_lock: threading.Lock = threading.Lock()
-        self._registration_lock: threading.Lock = threading.Lock()
-        self._atexit_registered: bool = False
-        # Default timeouts for known methods; unknown methods use a safe default
-        self._timeouts: dict[str, float] = {
-            "connect": 20.0,
-            "list_tools": 10.0,
-            "call_tool": 120.0,
-            "cleanup": 10.0,
-            "list_prompts": 10.0,
-            "get_prompt": 10.0,
-            "__aenter__": 15.0,
-            "__aexit__": 15.0,
-        }
-        # Server -> driver mapping (driver runs on background loop in a single task)
-        self._drivers: dict[Any, dict[str, Any]] = {}
-
-    def _resolve_method_timeout(self, server: Any, method_name: str) -> float:
-        """Resolve timeout for a method call, extending OAuth discovery waits only when needed."""
-        timeout = self._timeouts.get(method_name, 30.0)
-        if method_name != "list_tools":
-            return timeout
-
-        actual = getattr(server, "_server", server)
-        if _MCPServerOAuthClient is None or not isinstance(actual, _MCPServerOAuthClient):
-            return timeout
-
-        if _get_oauth_runtime_context is None:
-            return _OAUTH_LIST_TOOLS_TIMEOUT_SECONDS
-
-        runtime_context = _get_oauth_runtime_context()
-        runtime_timeout = getattr(runtime_context, "timeout", None) if runtime_context is not None else None
-        if isinstance(runtime_timeout, (int, float)) and runtime_timeout > 0:
-            return float(runtime_timeout) + _OAUTH_LIST_TOOLS_TIMEOUT_GRACE_SECONDS
-        return _OAUTH_LIST_TOOLS_TIMEOUT_SECONDS
-
-    def _ensure_driver(self, server: Any) -> None:
-        # Create a per-server driver task with a command queue if missing
-        real_server = getattr(server, "_server", server)
-        if real_server in self._drivers:
-            return
-        loop = self._ensure_bg_loop()
-        queue: asyncio.Queue = asyncio.Queue()
-        # Readiness event
-        ready_evt = threading.Event()
-        # Unwrap proxy to operate on the real server inside the driver (same task)
-
-        # Check if this is an OAuth client (two-phase auth)
-        is_oauth_client = _MCPServerOAuthClient is not None and isinstance(real_server, _MCPServerOAuthClient)
-
-        async def _driver():
-            # Connect once in this driver task to bind cancel scope and session
-            try:
-                if getattr(real_server, "session", None) is None and not getattr(
-                    real_server, "_discovery_session", None
-                ):
-                    server_name = getattr(real_server, "name", "<unnamed>")
-                    if is_oauth_client:
-                        # Two-phase auth: defer all connections to on-demand calls
-                        logger.info(
-                            f"Skipping eager discovery connect for OAuth server {server_name}; will connect on demand."
-                        )
-                    else:
-                        # Regular server: full connection
-                        logger.info(f"Connecting server {server_name}")
-                        await real_server.connect()
-            except Exception as conn_err:
-                # Log but don't crash - allow driver to start for retry/recovery
-                logger.error(f"Connection failed for {getattr(real_server, 'name', '<unnamed>')}: {conn_err}")
-            finally:
-                ready_evt.set()
-
-            while True:
-                cmd = await queue.get()
-                if cmd is None:
-                    continue
-                typ = cmd.get("type")
-                if typ == "call":
-                    method_name = cmd["method"]
-                    args = cmd.get("args", ())
-                    kwargs = cmd.get("kwargs", {})
-                    result_fut: Future = cmd["result_fut"]
-                    if _set_oauth_user_id is not None:
-                        _set_oauth_user_id(cast("str | None", cmd.get("oauth_user_id")))
-                    if _set_oauth_runtime_context is not None:
-                        _set_oauth_runtime_context(cmd.get("oauth_runtime_context"))
-                    try:
-                        method = getattr(real_server, method_name)
-                        res = await method(*args, **kwargs)
-                        result_fut.set_result(res)
-                    except BaseException as e:  # noqa: BLE001
-                        result_fut.set_exception(e)
-                    finally:
-                        if _set_oauth_runtime_context is not None:
-                            _set_oauth_runtime_context(None)
-                        if _set_oauth_user_id is not None:
-                            _set_oauth_user_id(None)
-                elif typ == "shutdown":
-                    result_fut: Future = cmd["result_fut"]
-                    try:
-                        cleanup = getattr(real_server, "cleanup", None)
-                        if callable(cleanup):
-                            cleanup_result = cleanup()
-                            if inspect.isawaitable(cleanup_result):
-                                await cleanup_result
-                        result_fut.set_result(True)
-                    except BaseException as e:  # noqa: BLE001
-                        result_fut.set_exception(e)
-                    break
-                elif typ == "force_stop":
-                    result_fut = cmd.get("result_fut")
-                    if isinstance(result_fut, Future) and not result_fut.done():
-                        result_fut.set_result(False)
-                    break
-
-        # Start driver
-        driver_future = asyncio.run_coroutine_threadsafe(_driver(), loop)
-        # Wait until driver has connected
-        if not ready_evt.wait(timeout=self._timeouts.get("connect", 20.0)):
-            # Handle timeout explicitly
-            raise TimeoutError(f"Server {getattr(server, 'name', '<unnamed>')} failed to connect within timeout")
-        # Track whether this driver created a session (regular or discovery)
-        has_session = getattr(real_server, "session", None) is not None
-        has_discovery = getattr(real_server, "_discovery_session", None) is not None
-        created_by_driver = has_session or has_discovery
-        self._drivers[real_server] = {
-            "queue": queue,
-            "real": real_server,
-            "created_by_driver": created_by_driver,
-            "driver_future": driver_future,
-        }
-
-    async def ensure_connected(self, server: Any) -> None:
-        # Ensure the per-server driver is running and connected
-        async with self._lock:
-            self._ensure_driver(server)
-
-    async def reconnect(self, server: Any) -> None:
-        """Force reconnection by clearing the existing driver and creating a new one.
-
-        Args:
-            server: The MCP server to reconnect (can be proxy or real server)
-        """
-        # Unwrap proxy to get real server
-        real_server = getattr(server, "_server", server)
-
-        async with self._lock:
-            # Clear the existing driver if present
-            if real_server in self._drivers:
-                server_name = getattr(real_server, "name", "<unnamed>")
-                logger.info(f"Clearing stale driver for {server_name}")
-                driver_state = self._drivers.pop(real_server)
-
-                # Try to cleanup the old driver gracefully
-                try:
-                    queue = driver_state.get("queue")
-                    if queue:
-                        # Send shutdown command
-                        from concurrent.futures import Future
-
-                        fut: Future = Future()
-                        queue.put_nowait({"type": "shutdown", "result_fut": fut})
-                        # Don't wait for it, just move on
-                except Exception:
-                    pass  # Ignore cleanup errors
-
-            # Clear session to force reconnection
-            if hasattr(real_server, "session"):
-                real_server.session = None
-
-            # Re-create the driver (this will reconnect)
-            self._ensure_driver(real_server)
-
-    async def connect_all(self) -> None:
-        for server in self._servers.values():
-            await self.ensure_connected(server)
-
-    async def shutdown(self) -> None:
-        """Cleanup all persistent servers and clear the registry."""
-        async with self._lock:
-            # Drive shutdown via driver queues to guarantee same-task cleanup
-            for _, state in list(self._drivers.items()):
-                queue: asyncio.Queue = state["queue"]
-                fut: Future = Future()
-
-                def _post(queue=queue, fut=fut):
-                    queue.put_nowait({"type": "shutdown", "result_fut": fut})
-
-                loop = self._ensure_bg_loop()
-                loop.call_soon_threadsafe(_post)
-                server_name = getattr(state.get("real"), "name", "<unnamed>")
-                try:
-                    fut.result(timeout=self._timeouts.get("cleanup", 10.0))
-                except TimeoutError:
-                    logger.warning(
-                        "Timed out waiting for MCP server '%s' cleanup; forcing shutdown",
-                        server_name,
-                    )
-
-                    def _force_stop(queue=queue, fut=fut):
-                        queue.put_nowait({"type": "force_stop", "result_fut": fut})
-
-                    loop.call_soon_threadsafe(_force_stop)
-                    try:
-                        fut.result(timeout=0.5)
-                    except TimeoutError:
-                        logger.warning(
-                            "Force-stop for MCP server '%s' did not complete in time",
-                            server_name,
-                        )
-                except Exception as exc:  # noqa: BLE001
-                    logger.warning(
-                        "Error during MCP server '%s' shutdown: %s",
-                        server_name,
-                        exc,
-                    )
-            self._drivers.clear()
-            self._servers.clear()
-            if self._bg_loop is not None:
-                try:
-                    self._bg_loop.call_soon_threadsafe(self._bg_loop.stop)
-                    if self._bg_thread is not None:
-                        self._bg_thread.join(timeout=2)
-                finally:
-                    self._bg_loop = None
-                    self._bg_thread = None
-
-    async def unregister_keys_ending_with(self, suffix: str) -> None:
-        """Cleanup and remove registered servers whose persistence keys end with suffix."""
-        if suffix == "":
-            return
-        async with self._lock:
-            matches = [key for key in self._servers if key.endswith(suffix)]
-            for key in matches:
-                server = self._servers.pop(key, None)
-                if server is not None:
-                    await self._shutdown_server_unlocked(server)
-
-    def register(self, server: Any, *, key: str | None = None) -> Any:
-        """Register (or reuse) a server by key and return the canonical instance."""
-        name = key or getattr(server, "name", None)
-        if not isinstance(name, str) or name == "":
-            # Do not persist unnamed servers
-            return server
-        existing = self._servers.get(name)
-        if existing is not None:
-            return existing
-        self._servers[name] = server
-        return server
-
-    def get(self, name: str) -> Any | None:
-        return self._servers.get(name)
-
-    def all(self) -> list[Any]:
-        return list(self._servers.values())
-
-    def update_oauth_cache_dir(self, cache_dir: Path | None) -> None:
-        """Update cache_dir for all OAuth-enabled servers registered with this manager."""
-        if not _OAUTH_AVAILABLE:
-            return
-        normalized = None
-        if cache_dir is not None:
-            normalized = cache_dir.expanduser()
-        for server in self._servers.values():
-            self._apply_cache_dir_to_server(server, normalized)
-        # Update drivers as well (LoopAffineAsyncProxy->real server)
-        for entry in self._drivers.values():
-            real_server = entry.get("real")
-            if real_server is not None:
-                self._apply_cache_dir_to_server(real_server, normalized)
-
-    def _apply_cache_dir_to_server(self, server: Any, cache_dir: Path | None) -> None:
-        """Internal helper to apply cache_dir to both configs and instantiated clients."""
-        if server is None:
-            return
-        actual = getattr(server, "_server", server)
-        if _MCPServerOAuth is not None and isinstance(actual, _MCPServerOAuth):
-            oauth_config = cast("MCPServerOAuth", actual)
-            apply_managed_oauth_cache_dir(oauth_config, cache_dir)
-            return
-        try:
-            from agency_swarm.mcp.oauth_client import MCPServerOAuthClient
-        except ImportError:  # pragma: no cover - optional dependency missing
-            return
-
-        if isinstance(actual, MCPServerOAuthClient):
-            config = actual.oauth_config
-            apply_managed_oauth_cache_dir(config, cache_dir)
-            oauth_provider = getattr(actual, "_oauth_provider", None)
-            storage = getattr(oauth_provider, "storage", None) if oauth_provider else None
-            if storage and hasattr(storage, "base_cache_dir") and cache_dir is not None:
-                storage.base_cache_dir = cache_dir
-
-    async def _shutdown_server_unlocked(self, server: Any) -> None:
-        real_server = getattr(server, "_server", server)
-        state = self._drivers.pop(real_server, None)
-        if state is not None:
-            queue: asyncio.Queue = state["queue"]
-            fut: Future = Future()
-            loop = self._ensure_bg_loop()
-            loop.call_soon_threadsafe(lambda: queue.put_nowait({"type": "shutdown", "result_fut": fut}))
-            running_loop = asyncio.get_running_loop()
-            try:
-                await running_loop.run_in_executor(None, fut.result, self._timeouts.get("cleanup", 10.0))
-            except TimeoutError:
-                driver_future = state.get("driver_future")
-                if isinstance(driver_future, Future):
-                    driver_future.cancel()
-            return
-
-        cleanup = getattr(real_server, "cleanup", None)
-        if callable(cleanup):
-            cleanup_result = cleanup()
-            if inspect.isawaitable(cleanup_result):
-                await cleanup_result
-
-    def _ensure_bg_loop(self) -> asyncio.AbstractEventLoop:
-        if self._bg_loop is not None:
-            return self._bg_loop
-        loop = asyncio.new_event_loop()
-
-        def _runner() -> None:
-            asyncio.set_event_loop(loop)
-            loop.run_forever()
-
-        thread = threading.Thread(target=_runner, name="mcp-persistence-loop", daemon=True)
-        thread.start()
-        self._bg_loop = loop
-        self._bg_thread = thread
-        return loop
-
-    def _submit_to_loop(self, coro: Any) -> Future:
-        loop = self._ensure_bg_loop()
-        return asyncio.run_coroutine_threadsafe(coro, loop)
-
-    def _submit_driver_call(self, server: Any, method: str, args: tuple[Any, ...], kwargs: dict[str, Any]) -> Future:
-        """Schedule a coroutine method call on the server's long-lived driver task."""
-        real_server = getattr(server, "_server", server)
-        self._ensure_driver(real_server)
-        state = self._drivers.get(real_server)
-        if state is None:
-            raise RuntimeError(f"Driver not initialized for server {getattr(real_server, 'name', '<unnamed>')}")
-
-        queue: asyncio.Queue = state["queue"]
-        fut: Future = Future()
-
-        def _post_call() -> None:
-            oauth_user_id = _get_oauth_user_id() if _get_oauth_user_id is not None else None
-            oauth_runtime_context = _get_oauth_runtime_context() if _get_oauth_runtime_context is not None else None
-            queue.put_nowait(
-                {
-                    "type": "call",
-                    "method": method,
-                    "args": args,
-                    "kwargs": kwargs,
-                    "oauth_user_id": oauth_user_id,
-                    "oauth_runtime_context": oauth_runtime_context,
-                    "result_fut": fut,
-                }
-            )
-
-        loop = self._ensure_bg_loop()
-        loop.call_soon_threadsafe(_post_call)
-        return fut
-
-    async def _await_future(self, fut: Future, timeout: float | None = None) -> Any:  # noqa: ANN401
-        loop = asyncio.get_running_loop()
-
-        def _get_result():
-            return fut.result(timeout=timeout)
-
-        return await loop.run_in_executor(None, _get_result)
-
-    def mark_atexit_registered(self) -> bool:
-        with self._registration_lock:
-            if self._atexit_registered:
-                return False
-            self._atexit_registered = True
-            return True
-
-    def shutdown_sync(self) -> None:
-        if not self._sync_shutdown_lock.acquire(blocking=False):
-            return
-        try:
-            try:
-                asyncio.run(self.shutdown())
-            except RuntimeError as exc:
-                message = str(exc)
-                if "asyncio.run() cannot be called from a running event loop" not in message:
-                    logger.warning("Error during persistent MCP manager shutdown: %s", exc)
-                    return
-                try:
-                    loop = asyncio.get_running_loop()
-                except RuntimeError as loop_error:
-                    logger.warning(
-                        "Error during persistent MCP manager shutdown: %s",
-                        loop_error,
-                    )
-                    return
-                loop.create_task(self.shutdown())
-            except Exception as exc:  # noqa: BLE001
-                logger.warning("Error during persistent MCP manager shutdown: %s", exc)
-        finally:
-            self._sync_shutdown_lock.release()
-
-
-class LoopAffineAsyncProxy:
-    """Generic proxy routing coroutine methods to the manager's background loop.
-
-    Avoids coupling to the concrete server implementation by dynamically proxying
-    any coroutine attribute via __getattr__.
-    """
-
-    def __init__(self, server: Any, manager: PersistentMCPServerManager) -> None:
-        self._server = server
-        self._manager = manager
-
-    async def __aenter__(self) -> Any:  # noqa: ANN401
-        target = getattr(self._server, "__aenter__", None)
-        if target is None:
-            raise TypeError(f"Server {self._server!r} does not support asynchronous context management")
-        timeout = self._manager._timeouts.get("__aenter__", 30.0)
-        if inspect.iscoroutinefunction(target):
-            fut = self._manager._submit_driver_call(self._server, "__aenter__", (), {})
-            return await self._manager._await_future(fut, timeout=timeout)
-        result = target()
-        if inspect.isawaitable(result):
-            fut = self._manager._submit_to_loop(result)
-            return await self._manager._await_future(fut, timeout=timeout)
-        return result
-
-    async def __aexit__(self, exc_type, exc, tb) -> Any:  # noqa: ANN001, ANN401
-        target = getattr(self._server, "__aexit__", None)
-        if target is None:
-            raise TypeError(f"Server {self._server!r} does not support asynchronous context management")
-        timeout = self._manager._timeouts.get("__aexit__", 30.0)
-        if inspect.iscoroutinefunction(target):
-            fut = self._manager._submit_driver_call(self._server, "__aexit__", (exc_type, exc, tb), {})
-            return await self._manager._await_future(fut, timeout=timeout)
-        result = target(exc_type, exc, tb)
-        if inspect.isawaitable(result):
-            fut = self._manager._submit_to_loop(result)
-            return await self._manager._await_future(fut, timeout=timeout)
-        return result
-
-    def __getattr__(self, name: str):  # noqa: ANN001
-        target = getattr(self._server, name)
-
-        if inspect.iscoroutinefunction(target):
-
-            async def _proxy(*args, **kwargs):  # noqa: ANN001
-                timeout = self._manager._resolve_method_timeout(self._server, name)
-                fut = self._manager._submit_driver_call(self._server, name, args, kwargs)
-                server_name = getattr(self._server, "name", "<unnamed>")
-                try:
-                    return await self._manager._await_future(fut, timeout=timeout)
-                except TimeoutError as exc:
-                    raise TimeoutError(
-                        f"MCP call '{name}' timed out after {timeout:.1f}s on server '{server_name}'"
-                    ) from exc
-                except asyncio.CancelledError as exc:
-                    current_task = asyncio.current_task()
-                    if current_task is not None and current_task.cancelling():
-                        raise
-                    raise RuntimeError(
-                        f"MCP call '{name}' was cancelled on server '{server_name}'. "
-                        "Check MCP server availability and OAuth configuration."
-                    ) from exc
-
-            return _proxy
-
-        return target
 
 
 default_mcp_manager = PersistentMCPServerManager()
-
-
-def _process_oauth_servers(agent: Any, servers: list[Any]) -> None:
-    """Process OAuth-enabled MCP servers and convert them to authenticated clients.
-
-    Args:
-        agent: The agent instance
-        servers: List of MCP server instances (may include MCPServerOAuth)
-    """
-    if not _OAUTH_AVAILABLE or _MCPServerOAuth is None:
-        return
-
-    # Import OAuth client here to avoid circular import
-    from agency_swarm.mcp.oauth_client import MCPServerOAuthClient
-
-    handler_factory = getattr(agent, "mcp_oauth_handler_factory", None)
-    factory: Callable[[str], OAuthHandlerMap] | None = None
-    if callable(handler_factory):
-        factory = cast("Callable[[str], OAuthHandlerMap]", handler_factory)
-
-    # Convert OAuth configs to OAuth clients
-    for i, srv in enumerate(list(servers)):
-        if not isinstance(srv, _MCPServerOAuth):
-            continue
-
-        oauth_srv = cast("MCPServerOAuth", srv)
-        logger.info(f"Creating OAuth client for MCP server: {oauth_srv.name}")
-
-        try:
-            client_id = oauth_srv.get_client_id_optional()
-            if client_id:
-                logger.info(f"OAuth configured for {oauth_srv.name} (client_id: {client_id[:8]}...)")
-
-            # Build handlers: config-level first. Factory remains a compatibility fallback.
-            server_handlers: OAuthHandlerMap = {}
-            if oauth_srv.redirect_handler:
-                server_handlers["redirect"] = oauth_srv.redirect_handler
-            if oauth_srv.callback_handler:
-                server_handlers["callback"] = oauth_srv.callback_handler
-
-            if factory is not None and not server_handlers:
-                new_handlers = factory(oauth_srv.name)
-                server_handlers.update(new_handlers)
-
-            handlers_arg = server_handlers if server_handlers else None
-            oauth_client = MCPServerOAuthClient(
-                oauth_srv,
-                handlers_arg,
-            )
-
-            # Replace config with actual client
-            servers[i] = oauth_client
-            logger.info(f"OAuth client created for {oauth_srv.name}")
-
-        except Exception:
-            logger.exception(f"Failed to create OAuth client for {oauth_srv.name}")
-            raise
-
-
-async def _authorize_hosted_mcp_tool(agent: Any, source_tool: Any, *, cache_dir: Path | None) -> Any | None:
-    """Return a request-local HostedMCPTool with an OAuth access token."""
-    if not _OAUTH_AVAILABLE or _MCPServerOAuth is None or _MCPServerOAuthClient is None:
-        return None
-    if _HostedMCPTool is None:
-        return None
-    source_tool_config = getattr(source_tool, "tool_config", None)
-    if not isinstance(source_tool_config, dict):
-        return None
-    server_label = source_tool_config.get("server_label")
-    server_url = source_tool_config.get("server_url")
-    if not isinstance(server_label, str) or server_label == "":
-        return None
-    if not isinstance(server_url, str) or server_url == "":
-        return None
-
-    active_tool = enable_hosted_mcp_tool_oauth(
-        _HostedMCPTool(
-            tool_config=cast(dict[str, Any], dict(source_tool_config)),
-            on_approval_request=getattr(source_tool, "on_approval_request", None),
-        )
-    )
-    setattr(active_tool, _HOSTED_MCP_OAUTH_ORIGINAL_TOOL_ATTR, source_tool)
-    active_tool_config = cast(dict[str, Any], active_tool.tool_config)
-    handler_factory = getattr(agent, "mcp_oauth_handler_factory", None)
-    factory: Callable[[str], OAuthHandlerMap] | None = None
-    if callable(handler_factory):
-        factory = cast("Callable[[str], OAuthHandlerMap]", handler_factory)
-
-    oauth_config_type = cast("type[MCPServerOAuth]", _MCPServerOAuth)
-    oauth_client_type = cast(type[Any], _MCPServerOAuthClient)
-    oauth_srv = oauth_config_type(url=server_url, name=server_label, use_env_credentials=False)
-    apply_managed_oauth_cache_dir(oauth_srv, cache_dir)
-    server_handlers: OAuthHandlerMap = {}
-    if factory is not None:
-        server_handlers.update(factory(server_label))
-    oauth_client = oauth_client_type(oauth_srv, server_handlers or None)
-
-    try:
-        await oauth_client.connect()
-        provider = getattr(oauth_client, "_oauth_provider", None)
-        if provider is None:
-            return None
-        tokens = await provider.context.storage.get_tokens()
-        if tokens is None or not getattr(tokens, "access_token", None):
-            return None
-        active_tool_config["authorization"] = tokens.access_token
-        return active_tool
-    finally:
-        await oauth_client.cleanup()
-
-
-def _stage_deferred_hosted_mcp_tools(agent: Any, *, cache_dir: Path | None) -> None:
-    """Withhold opted-in hosted tools until the model requests OAuth activation."""
-    tools = getattr(agent, "tools", None)
-    if not isinstance(tools, list) or getattr(agent, "_hosted_mcp_oauth_enabled", False) is not True:
-        return
-    if isinstance(getattr(agent, _HOSTED_MCP_OAUTH_ORIGINAL_TOOLS_ATTR, None), list):
-        return
-
-    candidates: dict[str, Any] = {}
-    for tool in tools:
-        if getattr(tool, "name", None) != "hosted_mcp" or not is_hosted_mcp_tool_oauth_enabled(tool):
-            continue
-        tool_config = getattr(tool, "tool_config", None)
-        if not isinstance(tool_config, dict) or tool_config.get("authorization") not in (None, ""):
-            continue
-        server_label = tool_config.get("server_label")
-        server_url = tool_config.get("server_url")
-        if not isinstance(server_label, str) or server_label == "":
-            continue
-        if not isinstance(server_url, str) or server_url == "":
-            continue
-        if server_label in candidates:
-            raise ValueError(f"Hosted MCP server name '{server_label}' must be unique for OAuth activation")
-        candidates[server_label] = tool
-    if not candidates:
-        return
-
-    local_servers = getattr(agent, "_oauth_mcp_servers", {})
-    local_names = set(local_servers) if isinstance(local_servers, dict) else set()
-    conflicts = local_names.intersection(candidates)
-    if conflicts:
-        conflict = sorted(conflicts)[0]
-        raise ValueError(f"MCP server name '{conflict}' is duplicated across local and hosted OAuth tools")
-
-    async def _activate_hosted_mcp(server_name: str) -> str | None:
-        selected = candidates.get(server_name)
-        if selected is None:
-            return None
-        active_tool = await _authorize_hosted_mcp_tool(agent, selected, cache_dir=cache_dir)
-        if active_tool is None:
-            return f"Failed to authenticate MCP server '{server_name}': no OAuth access token was returned"
-
-        current_tools = list(getattr(agent, "tools", []))
-        for index, tool in enumerate(current_tools):
-            if getattr(tool, _HOSTED_MCP_OAUTH_ORIGINAL_TOOL_ATTR, None) is selected:
-                current_tools[index] = active_tool
-                break
-        else:
-            current_tools.append(active_tool)
-        agent.tools = current_tools
-        return f"MCP server '{server_name}' is authenticated and its tools are enabled."
-
-    installer = getattr(agent, "_install_mcp_authentication_tool", None)
-    if not callable(installer):
-        return
-    original_tools = tools
-    activation_tool = next((tool for tool in tools if getattr(tool, "name", None) == "authenticate_mcp_server"), None)
-    original_schema = copy.deepcopy(getattr(activation_tool, "params_json_schema", None))
-    agent.tools = list(tools)
-    try:
-        installer(sorted(local_names.union(candidates)))
-    except BaseException:
-        agent.tools = original_tools
-        raise
-    if activation_tool is not None and isinstance(original_schema, dict):
-        request_schema = copy.deepcopy(activation_tool.params_json_schema)
-        activation_tool.params_json_schema = original_schema
-        request_activation = copy.copy(activation_tool)
-        request_activation.params_json_schema = request_schema
-        agent.tools = [request_activation if tool is activation_tool else tool for tool in agent.tools]
-    setattr(agent, _HOSTED_MCP_OAUTH_ORIGINAL_TOOLS_ATTR, original_tools)
-    setattr(agent, _HOSTED_MCP_OAUTH_ACTIVATION_HANDLER_ATTR, _activate_hosted_mcp)
-    withheld = {id(tool) for tool in candidates.values()}
-    agent.tools = [tool for tool in agent.tools if id(tool) not in withheld]
-
-
-def restore_hosted_mcp_oauth_tools(agency: Any) -> None:
-    """Restore HostedMCPTool definitions replaced by request-scoped OAuth injection."""
-    agents_map = getattr(agency, "agents", None)
-    if not isinstance(agents_map, dict):
-        return
-    for agent in agents_map.values():
-        tools = getattr(agent, "tools", None)
-        if not isinstance(tools, list):
-            continue
-        original_tools = getattr(agent, _HOSTED_MCP_OAUTH_ORIGINAL_TOOLS_ATTR, None)
-        if not isinstance(original_tools, list):
-            continue
-        agent.tools = original_tools
-        delattr(agent, _HOSTED_MCP_OAUTH_ORIGINAL_TOOLS_ATTR)
-        if hasattr(agent, _HOSTED_MCP_OAUTH_ACTIVATION_HANDLER_ATTR):
-            delattr(agent, _HOSTED_MCP_OAUTH_ACTIVATION_HANDLER_ATTR)
-
-
-def _sync_oauth_client_handlers(persistent: object, candidate: object) -> bool:
-    """Update cached OAuth client with per-request handlers from a new instance.
-
-    Returns False because sessions are no longer invalidated when handlers refresh.
-    """
-    if not _OAUTH_AVAILABLE or _MCPServerOAuthClient is None:
-        return False
-
-    existing_client = getattr(persistent, "_server", persistent)
-    new_client = getattr(candidate, "_server", candidate)
-
-    if not isinstance(existing_client, _MCPServerOAuthClient) or not isinstance(new_client, _MCPServerOAuthClient):
-        return False
-
-    server_handlers = _current_oauth_handlers(cast(Any, new_client))
-    if not server_handlers:
-        return False
-
-    client = cast(Any, existing_client)
-    if "redirect" in server_handlers:
-        client._redirect_handler = server_handlers["redirect"]
-    if "callback" in server_handlers:
-        client._callback_handler = server_handlers["callback"]
-    provider = getattr(client, "_oauth_provider", None)
-    provider_context = getattr(provider, "context", None) if provider is not None else None
-    if provider_context is not None:
-        if "redirect" in server_handlers:
-            provider_context.redirect_handler = server_handlers["redirect"]
-        if "callback" in server_handlers:
-            provider_context.callback_handler = server_handlers["callback"]
-    return False
 
 
 async def attach_persistent_mcp_servers(agency: Any) -> None:
